moved extract view to extract application
