+++ backend/infrastructures/apps/extract/views/__init__.py
@@ -1,0 +1,5 @@
+from .urls import extract_urlpatterns
+
+__all__ = [
+    "extract_urlpatterns",
+]
