+++ backend/infrastructures/apps/extract/views/serializers.py
@@ -1,0 +1,5 @@
+from rest_framework import serializers
+
+
+class ExtractDataSerializer(serializers.Serializer):
+    file_path = serializers.CharField(min_length=1)
