+++ backend/infrastructures/apps/extract/views/urls.py
@@ -1,0 +1,13 @@
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+
+from .views import ExtractViewSet
+
+
+extract_router = DefaultRouter()
+extract_router.register(prefix=r"", viewset=ExtractViewSet, basename="extract")
+
+
+extract_urlpatterns = [
+    path("", include(extract_router.urls)),
+]
