+++ backend/infrastructures/apps/load/views/serializers.py
@@ -1,10 +1,6 @@
 from rest_framework import serializers
 
 
-class ExtractDataSerializer(serializers.Serializer):
-    file_path = serializers.CharField(min_length=1)
-
-
 class OutputDataBaseSerializer(serializers.Serializer):
     #TODO discuss what to put here
     pass
