+++ backend/infrastructures/apps/load/views/urls.py
@@ -7,7 +7,6 @@ from .views import LoadViewSet
 load_router = DefaultRouter()
 load_router.register(prefix=r"", viewset=LoadViewSet, basename="load")
 
-
 load_urlpatterns = [
     path("", include(load_router.urls)),
 ]
